Backend built to serve arbitrary object
