+++ src/co2eq/plot_ietf_meetings.py
@@ -1,59 +1,6 @@
 from meeting import MeetingList
 from conf import CONF
 
-## building a meeting serie 
-meeting_list = [
-  { 
-    'name' : 'meeting_namet', 
-    'location' : { 
-      'country' : 'DE', 
-      'city' : 'BER',
-      'iata' : "BER",
-    },
-    'attendee_list': [                 
-      {  
-      "country": "NP" 
-      }, 
-      { 
-      "country": "JP"                       
-      },
-      {
-      "country": "NZ"
-      },
-      {
-      "country": "UK"
-      },
-      {
-      "country": "FR"
-      }
-    ]                                    
-  },
-  { 
-    'name' : 'meeting_namey', 
-    'location' : { 
-      'country' : 'DE', 
-      'city' : 'BER',
-      'iata' : "BER",
-    },
-    'attendee_list': [                 
-      {  
-      "country": "NP" 
-      }, 
-      { 
-      "country": "JP"                       
-      },
-      {
-      "country": "NZ"
-      },
-      {
-      "country": "UK"
-      },
-      {
-      "country": "FR"
-      }
-    ]                                    
-  }
-]
-
-ietf_meeting_list = MeetingList( name='Test', conf=CONF, meeting_list=meeting_list ) 
-ietf_meeting_list.plot_all()
+def get_meeting_list(data):
+  ietf_meeting_list = MeetingList( name=data['name'], conf=CONF, meeting_list=[ data ] ) 
+  ietf_meeting_list.plot_all()
